+++ main.py
@@ -3,7 +3,7 @@ from transport import Transport
 
 def main():
     display = Hyperpixel2r()
-    transport = Transport(display.screen)
+    transport = Transport(display.screen, display._exit, display._rawfb, display._updatefb())
 
     transport.run()  # Run the clock
