+++ transport.py
@@ -2,45 +2,26 @@
 import os
 import sys
 import signal
+import pygame
 
 from hsl import *
 from util import *
+# from hyperpixel2r import Hyperpixel2r
 
 # Credit to Pimoroni for the Hyperpixel2r class
-class Hyperpixel2r:
-    screen = None
-
-    def __init__(self):
-        self._init_display()
-
-        self.screen.fill((0, 0, 0))
-        if self._rawfb:
-            self._updatefb()
-        else:
-            pygame.display.update()
-
-        # For some reason the canvas needs a 7px vertical offset
-        # circular screens are weird...
-        self.center = (240, 247)
-        self._radius = 240
-
-        # Distance of hour marks from center
-        # self._marks = 220
-
-        self._running = False
-        self._origin = pygame.math.Vector2(*self.center)
-        self._clock = pygame.time.Clock()
-        self._colour = (255, 0, 255)
-
+class Transport:
+    # display = Hyperpixel2r()
+    def __init__(self, screen, _exit, _rawfb, _updatefb):
+        self.screen = screen
+        self._exit = _exit
+        self._rawfb = _rawfb
+        self._updatefb = _updatefb
         # Load the image, reduce the size of the tram icon and create img object
         # Credit for the icon source: https://www.flaticon.com/free-icons/train
-        size_single = (512 // 6, 358 // 6)
-        size_double = (1050 // 6, 367 // 6)
-        size_square = (512 // 10, 512 // 10)
-        self._img_double = load_and_scale_image("imgs/double-tram.png", size_double)
-        self._img_left = load_and_scale_image("imgs/single-tram-left.png", size_single)
-        self._img_right = load_and_scale_image("imgs/single-tram-right.png", size_single)
-        self._img_warning = load_and_scale_image("imgs/warning.png", size_square)
+        self._img_double = load_and_scale_image("imgs/double-tram.png", (1050 // 6, 367 // 6))
+        self._img_left = load_and_scale_image("imgs/single-tram-left.png", (512 // 6, 358 // 6))
+        self._img_right = load_and_scale_image("imgs/single-tram-right.png", (512 // 6, 358 // 6))
+        self._img_warning = load_and_scale_image("imgs/warning.png", (512 // 10, 512 // 10))
 
         # Initiate queue
         self.trip_status = None
@@ -58,83 +39,9 @@ class Hyperpixel2r:
         self.table_x = 40
         self.table_y = 300
         self.start_cell_x = 0
-
+        
+        self._clock = pygame.time.Clock()
         self._running = True
-
-    def _exit(self, sig, frame):
-        self._running = False
-        print("\nExiting!...\n")
-
-    def _init_display(self):
-        self._rawfb = False
-        # Based on "Python GUI in Linux frame buffer"
-        # http://www.karoltomala.com/blog/?p=679
-        DISPLAY = os.getenv("DISPLAY")
-        if DISPLAY:
-            print("Display: {0}".format(DISPLAY))
-
-        if os.getenv("SDL_VIDEODRIVER"):
-            print(
-                "Using driver specified by SDL_VIDEODRIVER: {}".format(
-                    os.getenv("SDL_VIDEODRIVER")
-                )
-            )
-            pygame.display.init()
-            size = (pygame.display.Info().current_w, pygame.display.Info().current_h)
-            if size == (480, 480):  # Fix for 480x480 mode offset
-                size = (640, 480)
-            self.screen = pygame.display.set_mode(
-                size,
-                pygame.FULLSCREEN
-                | pygame.DOUBLEBUF
-                | pygame.NOFRAME
-                | pygame.HWSURFACE,
-            )
-            return
-
-        else:
-            # Iterate through drivers and attempt to init/set_mode
-            for driver in ["rpi", "kmsdrm", "fbcon", "directfb", "svgalib"]:
-                os.putenv("SDL_VIDEODRIVER", driver)
-                try:
-                    pygame.display.init()
-                    size = (
-                        pygame.display.Info().current_w,
-                        pygame.display.Info().current_h,
-                    )
-                    if size == (480, 480):  # Fix for 480x480 mode offset
-                        size = (640, 480)
-                    self.screen = pygame.display.set_mode(
-                        size,
-                        pygame.FULLSCREEN
-                        | pygame.DOUBLEBUF
-                        | pygame.NOFRAME
-                        | pygame.HWSURFACE,
-                    )
-                    print(
-                        "Using driver: {0}, Framebuffer size: {1:d} x {2:d}".format(
-                            driver, *size
-                        )
-                    )
-                    return
-                except pygame.error as e:
-                    print('Driver "{0}" failed: {1}'.format(driver, e))
-                    continue
-                break
-
-        print("All SDL drivers failed, falling back to raw framebuffer access.")
-        self._rawfb = True
-        os.putenv("SDL_VIDEODRIVER", "dummy")
-        pygame.display.init()  # Need to init for .convert() to work
-        self.screen = pygame.Surface((480, 480))
-
-    def __del__(self):
-        "Destructor to make sure pygame shuts down, etc."
-
-    def _updatefb(self):
-        fbdev = os.getenv("SDL_FBDEV", "/dev/fb0")
-        with open(fbdev, "wb") as fb:
-            fb.write(self.screen.convert(16, 0).get_buffer())
     
     #--------------- Code for the clock starts here --------------------#
     def trip_table(self, data_queue, game_font, font_color, scroll_speed=0.15, clear_color=(0, 0, 0)):        
@@ -157,6 +64,7 @@ class Hyperpixel2r:
             # Only update if there is new data
             if self.trip_status != updated_data:
                 self.trip_status = updated_data
+        
         
         if self.trip_status is not None:
             # Clear screen before rendering new data
@@ -276,6 +184,7 @@ class Hyperpixel2r:
             self.screen.blit(self._img_double, (self.bottom_band_x, self.bottom_band_y))
 
     def run(self):
+        
         config = Transit_Config.get_config()
         trip_status = HSL_Trip_Update(config)
         service_message = HSL_Service_Alert(config)
@@ -304,7 +213,7 @@ class Hyperpixel2r:
             self.scrolling_bands(alert_queue, game_font, font_color)
 
             if self._rawfb:
-                self._updatefb()
+                self._updatefb
             else:
                 pygame.display.update()
                 pygame.event.pump()
@@ -316,6 +225,3 @@ class Hyperpixel2r:
         
         pygame.quit()
         sys.exit(0)
-
-display = Hyperpixel2r()
-display.run()
